Made a new limelight system with limeliteHelpers
